requirements.txt
biopython>=1.79
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
click>=8.0.0
pytest>=7.0.0
pytest-cov>=4.1.0
tqdm>=4.66.0